+++ backend/src/features/sales/dto.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
@@ -19,7 +20,7 @@ pub struct OrderItemResponse {
     pub order_id: Uuid,
     pub product_id: Uuid,
     pub quantity: i32,
-    pub unit_price: f64,
+    pub unit_price: Decimal,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -27,7 +28,7 @@ pub struct OrderResponse {
     pub id: Uuid,
     pub payment_method: String,
     pub status: String,
-    pub total_amount: f64,
+    pub total_amount: Decimal,
     pub items: Vec<OrderItemResponse>,
     pub created_at: DateTime<Utc>,
 }
